opsadmin/requirements.txt
amqp==1.4.9
ansible==2.2.0.0
anyjson==0.3.3
appnope==0.1.0
APScheduler==3.4.0
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.23.0
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
bcrypt==3.1.4
billiard==3.3.0.23
celery==3.1.25
celery-with-redis==3.0
certifi==2017.7.27.1
cffi==1.11.2
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.1
daphne==1.3.0
decorator==4.1.2
Django==1.11.6
django-jinja==2.4.1
django-redis==4.8.0
django-simple-captcha==0.5.5
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
Jinja2==2.10
kombu==3.0.37
ldap3==2.4
MarkupSafe==1.0
msgpack-python==0.4.8
olefile==0.44
paramiko==2.3.1
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.3.7
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.5.3
pykerberos==1.1.14
pyldap==2.4.45
PyMySQL==0.7.11
PyNaCl==1.1.2
python-memcached==1.58
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-kerberos==0.11.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
tzlocal==1.5.1
urllib3==1.22
wcwidth==0.1.7
zope.interface==4.4.3
